Moved gatherClues to Round.js
